+++ HISTORY.rst
@@ -3,12 +3,13 @@
 Release History
 ===============
 
-0.2.0b2 (unreleased)
+0.2.0b2 (2018-05-29)
 ++++++++++++++++++++
 
 - Added `namespace_suffix` to EventHubConfig() to support national clouds.
 - Added `device_id` attribute to EventData to support IoT Hub use cases.
 - Added message header to workaround service bug for PartitionKey support.
+- Updated uAMQP dependency to vRC1
 
 
 0.2.0b1 (2018-04-20)